# plan-ai/backend/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiResponse_null_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"TwentyManualConnectRequest": {
				"properties": {
					"baseUrl": {
						"type": "string",
						"description": "Root of the Twenty instance, e.g. \"https://crm.acme.com\"."
					},
					"apiKey": {
						"type": "string"
					}
				},
				"required": [
					"baseUrl",
					"apiKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TwentySummaryResponse": {
				"properties": {
					"connected": {
						"type": "boolean"
					},
					"baseUrl": {
						"type": "string"
					},
					"workspaceName": {
						"type": "string"
					}
				},
				"required": [
					"connected"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TwentySummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TwentySummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"TwentyCompanyItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"domainName": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TwentyCompanyItem-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/TwentyCompanyItem"
						},
						"type": "array",
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"TwentyPersonItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"companyId": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TwentyPersonItem-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/TwentyPersonItem"
						},
						"type": "array",
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"TwentyPushOutcome": {
				"type": "string",
				"enum": [
					"CREATED",
					"ALREADY_PUSHED",
					"DEDUPED"
				]
			},
			"PushTranscriptToTwentyResponse": {
				"properties": {
					"outcome": {
						"$ref": "#/components/schemas/TwentyPushOutcome"
					},
					"noteId": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"canonicalTranscriptId": {
						"type": "string",
						"description": "Present when outcome is DEDUPED — the recording that won."
					}
				},
				"required": [
					"outcome",
					"noteId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PushTranscriptToTwentyResponse-or-null_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/PushTranscriptToTwentyResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"PushTranscriptToTwentyRequest": {
				"properties": {
					"transcriptId": {
						"type": "string"
					},
					"companyId": {
						"type": "string"
					},
					"personIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"opportunityId": {
						"type": "string"
					},
					"forceSeparateNote": {
						"type": "boolean",
						"description": "Set by the UI when the user rejects an automatic \"same meeting\" match and\ninsists this really is a separate meeting deserving its own note."
					}
				},
				"required": [
					"transcriptId",
					"companyId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectTwentyLink": {
				"properties": {
					"projectId": {
						"type": "string"
					},
					"companyId": {
						"type": "string",
						"nullable": true
					},
					"companyName": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"projectId",
					"companyId",
					"companyName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ProjectTwentyLink-or-null_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ProjectTwentyLink"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"LinkProjectToTwentyCompanyRequest": {
				"description": "Links a Plan AI project to a company in Twenty.\n\nThis is what makes an unattended push possible. Every other integration can\nwork from a single workspace-wide default (a Jira project, a Trello board),\nbut a CRM note has to land on a DIFFERENT company per meeting — so the\ndestination is resolved per project, once, and every meeting recorded into\nthat project inherits it.",
				"properties": {
					"projectId": {
						"type": "string"
					},
					"companyId": {
						"type": "string",
						"nullable": true,
						"description": "Null unlinks the project (auto-push stops; manual push still works)."
					},
					"companyName": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"projectId",
					"companyId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrelloManualConnectRequest": {
				"properties": {
					"apiKey": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"apiKey",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TrelloSummaryResponse": {
				"properties": {
					"totalBoards": {
						"type": "number",
						"format": "double"
					},
					"latestBoards": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"totalBoards",
					"latestBoards"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TrelloSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TrelloSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"ApiResponse__success-boolean__": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"success": {
								"type": "boolean"
							}
						},
						"required": [
							"success"
						],
						"type": "object",
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"NotionSummaryResponse": {
				"properties": {
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"recentPages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"totalPages",
					"recentPages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_NotionSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/NotionSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"MicrosoftSummaryResponse": {
				"properties": {
					"isConnected": {
						"type": "boolean"
					},
					"userEmail": {
						"type": "string"
					}
				},
				"required": [
					"isConnected"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MicrosoftSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/MicrosoftSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"LinearManualConnectRequest": {
				"properties": {
					"apiKey": {
						"type": "string"
					}
				},
				"required": [
					"apiKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LinearSummaryResponse": {
				"properties": {
					"totalIssues": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"totalProjects": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"latestTeams": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"totalIssues",
					"totalProjects",
					"latestTeams"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_LinearSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/LinearSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"JiraManualConnectRequest": {
				"properties": {
					"siteUrl": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"apiToken": {
						"type": "string"
					}
				},
				"required": [
					"siteUrl",
					"email",
					"apiToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"JiraSummaryResponse": {
				"properties": {
					"totalIssues": {
						"type": "string",
						"nullable": true
					},
					"totalProjects": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"latestBoards": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"totalIssues",
					"totalProjects",
					"latestBoards"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_JiraSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/JiraSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"GoogleSummaryResponse": {
				"properties": {
					"isConnected": {
						"type": "boolean"
					},
					"userEmail": {
						"type": "string"
					}
				},
				"required": [
					"isConnected"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_GoogleSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/GoogleSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"AsanaManualConnectRequest": {
				"properties": {
					"personalAccessToken": {
						"type": "string"
					}
				},
				"required": [
					"personalAccessToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AsanaSummaryResponse": {
				"properties": {
					"totalTasks": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"totalProjects": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"totalTasks",
					"totalProjects"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_AsanaSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AsanaSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "4.3.3",
		"license": {
			"name": "BUSL-1.1"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/twenty/manual-connect": {
			"post": {
				"operationId": "ManualConnect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TwentyManualConnectRequest"
							}
						}
					}
				}
			}
		},
		"/api/twenty/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TwentySummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
		"/api/twenty/companies": {
			"get": {
				"operationId": "SearchCompanies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TwentyCompanyItem-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/twenty/people": {
			"get": {
				"operationId": "SearchPeople",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TwentyPersonItem-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/twenty/push-transcript": {
			"post": {
				"operationId": "PushTranscript",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PushTranscriptToTwentyResponse-or-null_"
								}
							}
						}
					},
					"400": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"404": {
						"description": "Transcript not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Push a meeting into Twenty as a note linked to the company (and people).\nDeduplicated server-side: if a teammate already pushed the same meeting,\nthis returns their note instead of creating a second one.",
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PushTranscriptToTwentyRequest"
							}
						}
					}
				}
			}
		},
		"/api/twenty/link-project": {
			"post": {
				"operationId": "LinkProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ProjectTwentyLink-or-null_"
								}
							}
						}
					},
					"404": {
						"description": "Project not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Links a project to a Twenty company so meetings recorded into it can be\npushed unattended (the \"Send to Twenty\" toggle in the recorder / mobile).",
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LinkProjectToTwentyCompanyRequest"
							}
						}
					}
				}
			}
		},
		"/api/trello/manual-connect": {
			"post": {
				"operationId": "ManualConnect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TrelloManualConnectRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/trello/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TrelloSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/notion/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_NotionSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/microsoft/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MicrosoftSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/linear/manual-connect": {
			"post": {
				"operationId": "ManualConnect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LinearManualConnectRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/linear/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_LinearSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/jira/manual-connect": {
			"post": {
				"operationId": "ManualConnect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__success-boolean__"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/JiraManualConnectRequest"
							}
						}
					}
				}
			}
		},
		"/api/jira/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_JiraSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/google/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_GoogleSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/asana/manual-connect": {
			"post": {
				"operationId": "ManualConnect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__success-boolean__"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AsanaManualConnectRequest"
							}
						}
					}
				}
			}
		},
		"/api/asana/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AsanaSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}